+++ src/Http/Controllers/TestController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Wotta\CommandExtender\Http\Controllers;
+
+class TestController extends \Illuminate\Routing\Controller
+{
+    public function index()
+    {
+        return 'hello';
+    }
+}