+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to `command-extender` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.0.0
 
 - initial release
+- Added `--action` for the `route:list` command.
+- Added `--open` option for `route:list` command.